backend/requirements.txt
langchain
groq
openai
python-dotenv
pytesseract
pymupdf
pytube
pinecone-client
langchain-pinecone
langchain-community
langchain-openai
langchain-groq
beautifulsoup4
httpx
tldextract
numpy
pillow
youtube-transcript-api
yt-dlp
langchain-text-splitters
html2text
langchain-core
urllib3
aiohttp
asyncio
unstructured
pdf2image
ocrmypdf
langchain_experimental
tiktoken
certifi
requests
pyOpenSSL
